wall-paint-visualizer-agent/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.0
diffusers==0.21.4
transformers==4.33.2
accelerate==0.23.0
langchain==0.0.311
langchain-community==0.0.10
torch==2.0.1
numpy==1.25.2
pydantic==2.4.2
httpx==0.24.1